requirements.txt
google-generativeai==0.7.2
httpx[http2]==0.27.2
jupyter==1.1.1
langchain==0.3.4
langchain-community==0.3.3
langchain-google-genai==2.0.1
matplotlib==3.9.2
pandas==2.2.3
pypdf==5.0.1
python-dotenv==1.0.1
scikit-learn==1.5.2
seaborn==0.13.2
semanticscholar==0.8.4